refactor: EventHandler를 이용하여 의존성 단방향으로 변경
